Storage: use DataSource in test online resize (#2300)
* use RHEL Datasource into test_online_resize.py test

* Fix PRs comments after review.

* Pre-commit run changes applied.

* Change PRs review comments.

* Change create_rhel_dv location and name.

* Fix some suggestions after PR's review

* Fix issues requested after PR reviews.

* Apply requested changes from PR review.

* Apply requested changes from PR review.

* Change naming in function from PR review.
